blogwebsite/requirements.txt
Django==4.2.4
psycopg2-binary==2.9.7
djangorestframework==3.14
